fixup! refactor(ivy): create Injector interface; remove dependency on Ivy
