+++ packages/core/src/di/r3_injector.ts
@@ -13,8 +13,7 @@ import {stringify} from '../util/stringify';
 import {resolveForwardRef} from './forward_ref';
 import {INJECTOR, Injector, NullInjector, THROW_IF_NOT_FOUND, USE_VALUE} from './injector';
 import {inject, injectArgs, setCurrentInjector} from './injector_compatibility';
-import {InjectFlags, InjectableDef, InjectableType, InjectionToken, InjectorType, InjectorTypeWithProviders, getInjectableDef, getInjectorDef} from './interface';
-import {ClassProvider, ConstructorProvider, ExistingProvider, FactoryProvider, Provider, StaticClassProvider, StaticProvider, TypeProvider, ValueProvider} from './interface/provider';
+import {ClassProvider, ConstructorProvider, ExistingProvider, FactoryProvider, InjectFlags, InjectableDef, InjectableType, InjectionToken, InjectorType, InjectorTypeWithProviders, Provider, StaticClassProvider, StaticProvider, TypeProvider, ValueProvider, getInjectableDef, getInjectorDef} from './interface';
 import {APP_ROOT} from './scope';
